# improve performance of id creation