… spectator-api/src/main/java/com/netflix/spectator/api/TagList.java
… spectator-api/src/test/java/com/netflix/spectator/api/ExtendedRegistryTest.java
… spectator-api/src/test/java/com/netflix/spectator/api/NoopIdTest.java
… spectator-api/src/test/java/com/netflix/spectator/api/NoopRegistryTest.java
… spectator-api/src/test/java/com/netflix/spectator/api/TagListTest.java
… spectator-api/src/test/java/com/netflix/spectator/api/UtilsTest.java