Switches internal tag list to an immutable tag set based
on a sorted array. While still considerably slower than
the simple prepend it improves on the previous behavior
and doesn't change any of the public APIs or expectations
around how the id behaves.

**Before**

```
Benchmark                   Mode  Cnt          Score         Error  Units
Ids.append1                thrpt   50    2,469,959.918 ±   51911.595  ops/s
Ids.append2                thrpt   50    1,169,621.422 ±   16694.934  ops/s
Ids.append4                thrpt   50      533,355.365 ±    6421.775  ops/s
Ids.baseline               thrpt   50   16,789,863.364 ±  224794.375  ops/s
Ids.justName               thrpt   50  159,751,023.136 ± 1377535.064  ops/s
Ids.withTag                thrpt   50      297,760.744 ±    2113.190  ops/s
Ids.withTagsMap            thrpt   50    1,775,862.197 ±   15440.667  ops/s
Ids.withTagsVararg         thrpt   50      289,581.182 ±    4740.400  ops/s
```

**After**

```
Benchmark                   Mode  Cnt          Score         Error  Units
Ids.append1                thrpt   50   17,966,856.652 ±  485452.845  ops/s
Ids.append2                thrpt   50    4,366,074.897 ±   82429.423  ops/s
Ids.append4                thrpt   50    3,323,952.076 ±   26348.669  ops/s
Ids.baseline               thrpt   50   17,107,693.322 ±  231479.939  ops/s
Ids.justName               thrpt   50  162,614,450.489 ± 1696642.250  ops/s
Ids.withTag                thrpt   50    1,392,055.132 ±   13903.741  ops/s
Ids.withTagsMap            thrpt   50    1,937,955.362 ±   17493.272  ops/s
Ids.withTagsVararg         thrpt   50    2,041,489.037 ±   20497.897  ops/s
```

**Summary**

| Test            | Before  | After   |
|-----------------|---------|---------|
| justName        |   1.0   |   1.0   |
| withTag         |   1.0   | **4.7** |
| withTagsMap     |   1.0   |   1.1   |
| withTagsVararg  |   1.0   | **7.0** |
| append1         |   1.0   | **7.3** |
| append2         |   1.0   | **3.7** |
| append4         |   1.0   | **6.2** |