@brharrington 2016-04-25
Added some of @dmuino 's comments. We'll merge and follow up on some of the other items in laters PRs.